+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fix: empty table cells sometimes caused the document to become corrupted in Markdown to DOCX converter
 - Fix: don't reset numbering in the target document when appending Markdown to an existing DOCX document (issue #12)
 - Improvements to styles and lists handling when appending Markdown to an existing DOCX document (issue #12)
+- Heading styles used in Markdown to DOCX converter now inherit from default Word heading styles, so that collapsing/expanding is available.
+- Fix: SVG images are now preserved in Markdown to DOCX renderer
 
 ## 0.13.1 - 2025.08.31